requirements.txt
Django==3.2.9
Pillow==8.4.0

django-environ==0.8.1
psycopg2-binary==2.9.2
gunicorn==20.1.0

django-bootstrap-v5==1.0.7
django-extensions==3.1.5
django-tagulous==1.3.0
Faker==10.0.0

thefuzz==0.19.0
tqdm==4.62.3

django-debug-toolbar
